refactor: Lines 패키지 위치 수정 및 Line.of 메서드 명 수정
