+++ src/test/java/nextstep/ladder/LinesTest.java
@@ -1,6 +1,7 @@
 package nextstep.ladder;
 
 import nextstep.ladder.domain.Line;
+import nextstep.ladder.domain.Lines;
 import nextstep.ladder.domain.Point;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
